requirements.txt
click
pillow
boto3
pandas
progress